pyproject.toml
[project]
name = "cf-pipeline"
version = "0.1.0"
description = "Research-grade hybrid CF pipeline on MovieLens 1M"
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.26",
  "pandas>=2.2",
  "pyarrow>=15",
  "scipy>=1.12",
  "scikit-learn>=1.4",
  "torch>=2.3",
  "lightgbm>=4.3",
  "tqdm>=4.66",
  "hydra-core>=1.3",
  "omegaconf>=2.3",
  "sentence-transformers>=2.7",
  "faiss-cpu>=1.8",
  "rank-bm25>=0.2.2",
  "transformers>=4.43",
  "peft>=0.11",
  "accelerate>=0.30",
  "datasets>=2.19",
  "bitsandbytes>=0.43",
  "optuna>=3.6",
  "matplotlib>=3.8",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov>=5.0", "ruff>=0.4"]

[tool.uv]
package = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"
markers = [
  "integration: requires processed data on disk (run scripts/prepare_data.py first)",
  "gpu: requires GPU + gemma-2-9b-it downloaded",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu124", marker = "sys_platform == 'win32'" },
]
